feat:fetch requests optimization test - 1
