+++ package.json
@@ -77,7 +77,6 @@
   },
   "lint-staged": {
     "src/**/*.{ts,tsx}": [
-      "yarn lint:fix",
       "yarn prettier:fix"
     ]
   }
