+++ src/app/layout.tsx
@@ -1,6 +1,3 @@
-// export const fetchCache = 'force-no-store'
-// export const revalidate = 0
-
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import './globals.css'
@@ -9,6 +6,25 @@ import { ProviderWrapper } from '@/redux/ProviderWrapper'
 import './tapwrite.css'
 import { InterrupCmdK } from '@/hoc/Interrupt_CmdK'
 import { ProgressLoad } from '@/hoc/ProgressLoader'
+import { headers } from 'next/headers'
+import { WorkflowStateResponse } from '@/types/dto/workflowStates.dto'
+import { apiUrl } from '@/config'
+import { ClientSideStateUpdate } from '@/hoc/ClientSideStateUpdate'
+import { TaskResponse } from '@/types/dto/tasks.dto'
+import { IAssignee, ITemplate } from '@/types/interfaces'
+import { addTypeToAssignee } from '@/utils/addTypeToAssignee'
+import ClientError from '@/components/clientError'
+import { Token, TokenSchema } from '@/types/common'
+import { CopilotAPI } from '@/utils/CopilotAPI'
+import { CreateAttachmentRequest } from '@/types/dto/attachments.dto'
+import { CreateViewSettingsDTO } from '@/types/dto/viewSettings.dto'
+import { createMultipleAttachments, getSignedUrlUpload } from '@/app/actions'
+import { ModalNewTaskForm } from './ui/Modal_NewTaskForm'
+import { MAX_FETCH_ASSIGNEE_COUNT } from '@/constants/users'
+import { RealTime } from '@/hoc/RealTime'
+import { redirectIfTaskCta } from '@/utils/redirect'
+import { sortTaskByDescendingOrder } from '@/utils/sortTask'
+import { z } from 'zod'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -17,14 +33,91 @@ export const metadata: Metadata = {
   description: 'A comprehensive task app',
 }
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+async function getAllWorkflowStates(token: string): Promise<WorkflowStateResponse[]> {
+  const res = await fetch(`${apiUrl}/api/workflow-states?token=${token}`, {
+    next: { tags: ['getAllWorkflowStates'] },
+  })
+
+  const data = await res.json()
+
+  return data.workflowStates
+}
+
+async function getAllTasks(token: string): Promise<TaskResponse[]> {
+  const res = await fetch(`${apiUrl}/api/tasks?token=${token}`, {
+    next: { tags: ['getTasks'] },
+  })
+
+  const data = await res.json()
+
+  return sortTaskByDescendingOrder(data.tasks)
+}
+
+async function getTokenPayload(token: string): Promise<Token> {
+  const copilotClient = new CopilotAPI(token)
+  const payload = TokenSchema.parse(await copilotClient.getTokenPayload())
+  return payload as Token
+}
+
+async function getAssigneeList(token: string): Promise<IAssignee> {
+  const res = await fetch(`${apiUrl}/api/users?token=${token}&limit=${MAX_FETCH_ASSIGNEE_COUNT}`, {
+    next: { tags: ['getAssigneeList'] },
+  })
+
+  const data = await res.json()
+
+  return data.users
+}
+
+async function getViewSettings(token: string): Promise<CreateViewSettingsDTO> {
+  const res = await fetch(`${apiUrl}/api/view-settings?token=${token}`, {
+    next: { tags: ['getViewSettings'] },
+  })
+  const data = await res.json()
+
+  return data
+}
+
+async function getAllTemplates(token: string): Promise<ITemplate[]> {
+  const res = await fetch(`${apiUrl}/api/tasks/templates?token=${token}`, {})
+
+  const templates = await res.json()
+
+  return templates.data
+}
+
+export default async function RootLayout({ children }: { children: React.ReactNode }) {
+  const headersList = headers()
+  const token = z.string().parse(headersList.get('clientToken'))
+
+  const [workflowStates, tasks, assignee, viewSettings, tokenPayload, templates] = await Promise.all([
+    getAllWorkflowStates(token),
+    getAllTasks(token),
+    addTypeToAssignee(await getAssigneeList(token)),
+    getViewSettings(token),
+    getTokenPayload(token),
+    getAllTemplates(token),
+  ])
+
   return (
     <html lang="en">
       <body className={inter.className}>
         <ProgressLoad />
         <InterrupCmdK>
           <ProviderWrapper>
-            <ThemeRegistry options={{ key: 'mui' }}>{children}</ThemeRegistry>
+            <ClientSideStateUpdate
+              workflowStates={workflowStates}
+              tasks={tasks}
+              token={token}
+              assignee={assignee}
+              viewSettings={viewSettings}
+              tokenPayload={tokenPayload}
+              templates={templates}
+            >
+              <RealTime>
+                <ThemeRegistry options={{ key: 'mui' }}>{children}</ThemeRegistry>
+              </RealTime>
+            </ClientSideStateUpdate>
           </ProviderWrapper>
         </InterrupCmdK>
       </body>
